修改

+++ src/main/java/org/fms/eis/webapp/vo/PCalcTaskVO.java
@@ -17,21 +17,21 @@ public class PCalcTaskVO extends ManagerParamVO {
 
     private Long id;    //任务标识
     private String name;    //名称
-    private BigDecimal tasktypeid;    //任务类型
+    private Integer tasktypeid;    //任务类型
     private String objtype;    //对象类型(计算方案模板类型【TPL_TYPE】)
-    private BigDecimal schemeId;    //计算方案
+    private Integer schemeId;    //计算方案
     private String primaryFlag;    //主方案标志
     private String dataSrc;    //数据源(日冻结、月冻结、终端冻结、曲线等)
     private String gltsFlag;    //功率、费率标识(每8位为一组，表示8个功率，每一组中的每一位代表8个费率)
-    private BigDecimal tsSchemeid;    //后台分时方案(P_SYS_RATE_SHEME)
+    private Long tsSchemeid;    //后台分时方案(P_SYS_RATE_SHEME)
     private String triggerFlag;    //是否触发对象(01-是 02-否)
     private String qCalcFlag;    //无功计算方式(废弃：直接设置到点上)
     private String priority;    //优先级
     private String genedatetime;    //任务产生基准时间(05 00:00 日 小时:分钟)
-    private BigDecimal genespanvalue;    //任务间隔值(>=1)
+    private Integer genespanvalue;    //任务间隔值(>=1)
     private String genespantype;    //任务间隔类型
     private String status;    //任务状态(0---停止 1---使用)
-    private BigDecimal weight;//排序
+    private Integer weight;//排序
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date taskst;    //任务产生开始时间
@@ -58,11 +58,11 @@ public class PCalcTaskVO extends ManagerParamVO {
         this.name = name;
     }
 
-    public BigDecimal getTasktypeid() {
+    public Integer getTasktypeid() {
         return tasktypeid;
     }
 
-    public void setTasktypeid(BigDecimal tasktypeid) {
+    public void setTasktypeid(Integer tasktypeid) {
         this.tasktypeid = tasktypeid;
     }
 
@@ -74,11 +74,11 @@ public class PCalcTaskVO extends ManagerParamVO {
         this.objtype = objtype;
     }
 
-    public BigDecimal getSchemeId() {
+    public Integer getSchemeId() {
         return schemeId;
     }
 
-    public void setSchemeId(BigDecimal schemeId) {
+    public void setSchemeId(Integer schemeId) {
         this.schemeId = schemeId;
     }
 
@@ -106,11 +106,11 @@ public class PCalcTaskVO extends ManagerParamVO {
         this.gltsFlag = gltsFlag;
     }
 
-    public BigDecimal getTsSchemeid() {
+    public Long getTsSchemeid() {
         return tsSchemeid;
     }
 
-    public void setTsSchemeid(BigDecimal tsSchemeid) {
+    public void setTsSchemeid(Long tsSchemeid) {
         this.tsSchemeid = tsSchemeid;
     }
 
@@ -146,11 +146,11 @@ public class PCalcTaskVO extends ManagerParamVO {
         this.genedatetime = genedatetime;
     }
 
-    public BigDecimal getGenespanvalue() {
+    public Integer getGenespanvalue() {
         return genespanvalue;
     }
 
-    public void setGenespanvalue(BigDecimal genespanvalue) {
+    public void setGenespanvalue(Integer genespanvalue) {
         this.genespanvalue = genespanvalue;
     }
 
@@ -210,11 +210,11 @@ public class PCalcTaskVO extends ManagerParamVO {
         this.lastModifyTime = lastModifyTime;
     }
 
-    public BigDecimal getWeight() {
+    public Integer getWeight() {
         return weight;
     }
 
-    public void setWeight(BigDecimal weight) {
+    public void setWeight(Integer weight) {
         this.weight = weight;
     }
